Update Positions
Still not functional, but at least it is loading everything correctly -- tomorrow I am going to start a new branch and scrap the table to try displaying cards - then I'll go from there.

 Relevant Errors being thrown are:

React.createElement: type is invalid -- expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Uncaught Invariant Violation: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

These each get thrown twice, presumably per attempted addition of the table row from PositionItem -- based on the errors, the assumption is that there is a type mismatch, or the table is unable to display things.  In any case, I will reconstruct the component and display tomorrow, if not tonight.